+++ _clear_ctx.ts
@@ -1,7 +1,7 @@
 /**
  * Returns obj with keys in `key_a1` having `clear_value`.
  */
-export function _clear_ctx<I extends unknown>(
+export function _clear_ctx<I extends unknown = unknown>(
 	key_a1: string[],
 	clear_value: I
 ) {