+++ _pick_value_a1.d.ts
@@ -1,5 +1,5 @@
 /**
  * Returns an array of objects with [pick](#pick) applied.
  */
-export declare function _pick_value_a1<I extends unknown>(obj: Record<string, I>, ...key_a1: string[]): I[] | undefined;
+export declare function _pick_value_a1<I extends unknown = unknown>(obj: Record<string, I>, ...key_a1: string[]): I[] | undefined;
 export declare const _a1__value__pick: typeof _pick_value_a1;