+++ assign.d.ts
@@ -1,4 +1,4 @@
 /**
  * Object.assign
  */
-export declare function assign<I extends object>(obj: I, ...arg_a1: unknown[]): I;
+export declare function assign<I extends object = object>(obj: I, ...arg_a1: unknown[]): I;