+++ call_assign.d.ts
@@ -1,6 +1,6 @@
 /**
  * Assigns function calls into obj
  */
-export declare function call_assign<I extends unknown>(obj: Record<string, I>, ...assign_fn_h_a1: Record<string, assign_fn_type<I>>[]): Record<string, I>;
+export declare function call_assign<I extends unknown = unknown>(obj: Record<string, I>, ...assign_fn_h_a1: Record<string, assign_fn_type<I>>[]): Record<string, I>;
 export declare const assign__call: typeof call_assign;
-export declare type assign_fn_type<I extends unknown> = (value: I, obj?: Record<string, I>, key?: string) => I;
+export declare type assign_fn_type<I extends unknown = unknown> = (value: I, obj?: Record<string, I>, key?: string) => I;