+++ clone.d.ts
@@ -1,4 +1,4 @@
 /**
  * Assigns arguments to new object
  */
-export declare function clone<I extends object>(...arg_a1: Partial<I>[]): I;
+export declare function clone<I extends object = object>(...arg_a1: Partial<I>[]): I;