+++ defaults.d.ts
@@ -1,4 +1,4 @@
 /**
  * Returns the obj with default values. If `obj[key] == null`, use `default[key]`
  */
-export declare function defaults<I extends object>(obj: I, ...defaults_a1: Partial<I>[]): I;
+export declare function defaults<I extends object = object>(obj: I, ...defaults_a1: Partial<I>[]): I;