+++ some.d.ts
@@ -5,4 +5,4 @@ import type { maybe } from '@ctx-core/function';
  * some({foo: 9, bar: 10}, (value, key) => value === 10) // returns true
  * some({baz: 11, quux: 12}, (value, key) => value === 10) // returns false
  */
-export declare function some<I extends unknown>(obj: Record<string, I>, compare_some: (obj: I, key: string) => unknown): maybe<boolean>;
+export declare function some<I extends unknown = unknown>(obj: Record<string, I>, compare_some: (obj: I, key: string) => unknown): maybe<boolean>;