+++ unpick.ts
@@ -3,7 +3,7 @@ import type { maybe } from '@ctx-core/function'
 /**
  * Does not include `key_a1` from `obj`
  */
-export function unpick<I extends unknown>(
+export function unpick<I extends unknown = unknown>(
 	obj:Record<string, I>,
 	...key_a1:string[]
 ):maybe<Record<string, I>> {