+++ unpick_keys.ts
@@ -3,7 +3,7 @@ import { keys } from './keys'
 /**
  * Does not include keys on `keys_o` from `rec`
  */
-export function unpick_keys<I extends unknown>(
+export function unpick_keys<I extends unknown = unknown>(
 	rec:Record<string, I>,
 	keys_o:Record<string, unknown>
 ) {